add email registered error widget
